+++ src/Models/Telemetry/Telemetry.hs
@@ -1452,7 +1452,7 @@ getProjectStatsBySpanType projectId start end = do
       |]
 
 
-getEndpointStats :: DB es => Projects.ProjectId -> UTCTime -> UTCTime -> Eff es [(Text, Text, Text, Int, Int)]
+getEndpointStats :: DB es => Projects.ProjectId -> UTCTime -> UTCTime -> Eff es [(Text, Text, Text, Int64, Int64)]
 getEndpointStats projectId start end = do
   Hasql.interp
     [HI.sql|
@@ -1460,7 +1460,7 @@ SELECT
     COALESCE(attributes___server___address, attributes->'net'->'host'->>'name', attributes->'http'->>'host', NULLIF(split_part(split_part(split_part(attributes___url___full, '://', 2), '/', 1), ':', 1), ''), resource___service___name, '') AS host,
     COALESCE(attributes___http___request___method, 'GET') AS method,
     COALESCE(attributes___url___path, '') AS url_path,
-    CAST(ROUND(AVG(COALESCE(duration, 0))) AS INT) AS average_duration,
+    CAST(ROUND(AVG(COALESCE(duration, 0))) AS BIGINT) AS average_duration,
     COUNT(*)::bigint AS request_count
 FROM otel_logs_and_spans
 WHERE
